requirements.txt
fastapi
uvicorn
opencv-python
numpy
paho-mqtt==1.6.1
Pillow
python-multipart
setuptools
bcrypt
python-jose[cryptography]
passlib